Create run_timer.py
This is a script that runs a timer in Python. The script asks for integer values for the year, month, day, time in hours, minutes, and seconds. It also executes the current date and time with the datetime module to see if it passed the given time that was specified.